Initial Commit
Started this project two days ago just to test as a demo, it looked to be turning out well! Seems like a fun thing to make into a repo.

I was debating forking some of these already existing projects out there, and upstreaming the changes to those, rather than making my own new project. I have a few new ideas how I want to implement it though, so upstreaming a whole new codebase to one of the older projects didn't seem to make as much sense after I thought about it. Especially if I were to totally change the API for it, too, which I think I will make from scratch. I think having an ESM-based syntax for these kinds of things work really well, and with the more recently added native Node fetch, also! Now you can just call `fetch()` like in the browser, super awesome in working with interface types and Promises. I really like that setup.

I think the goal for this project will be to make an ESM API with Promises and TypeScript definitions for the Mojang API, which will provide all of the same base features that the older libraries provide. This will be super cool to have for both use in the browser, and Node. Just like all of my other projects it seems! You can probably see a trend of things I like at this point, haha.

https://ourcodeworld.com/articles/read/1293/how-to-retrieve-the-skin-of-a-minecraft-user-from-mojang-using-python-3 (Found this article way back in 2020 I think, it was too complicated for me to figure out how to do, I was interested in trying to get it to work eventually. Glad I found it recently again this week, now I'm to the level where I have the different coding tools where it makes sense, and is doable now! Awesome. Super happy when reflective moments like that happen every once in a while. Another one of those was like when I redid Flatlands with the Canvas API)
https://wiki.vg/Mojang_API (Super helpful, am going to use this a bunch to get all of the types ready to define the API. TypeScript is absolutely killer for this kind of stuff, it helps so much!)
https://github.com/minecrafter/mojang-api (This was the one I was going to fork and makes updates/rewrites too)
https://github.com/mallusrgreatv2/all-minecraft
https://github.com/astra137/node-mojang

And the API goals for this project are to make things functional, I think. I like having dedicated functions for doing different things, and accessing each part of the API. Some of those other projects use classes and such, but I want to work with the API on a more object and property-focused setup. Maybe this could be a fun project to turn into a CLI tool, too who knows?? I still want to learn about setting up a JS project to use with the CLI directly, like as a dedicated globally-installed command. Maybe even bundling it with nexe or pkg, those seem like neat projects.