Demo Script
This project is still in the very early idea stages, so the code quality is sub-par as of now. Learning how to do some things a few different ways. Not sure what this can be used for yet, either. Always wanted to make my own skin downloading site, so maybe this could be for that?

Some of these changes were going to be made the other day also, but I ended up accidentally `git checkout .`-ing my small few edits, waht a dehrp.

Oh yeah, it was also because I couldn't come up with a name for the project. I wanted to find something clever, since `mojang-api`, `mojang`, and `mine-api` are all already taken. I was going to do `poisonous-potato`, but that was too random to remember. It doesn't really connect with the library at all either. I think changing back and forth on names is what broke the changes I made actually, now that I remember. I was also gonna go with `golden-apple`, but wasn't in love with the name. Saying it (lol) out loud again (by that, I mean typing it) though, it actually might be neat afterall. We'll see (I don't like that phrase, but I guess I'm using it here, aaagh)!